- Moved Gallery Module to (modules)/gallery.
