+++ src/app/admin/(modules)/gallery/[imageId]/loading.tsx
@@ -1,0 +1,5 @@
+import ShowWrapperLoading from "@/app/admin/(modules)/gallery/[imageId]/_components/Loading/ShowWrapperLoading";
+
+const AdminGalleryItemShowLoading = () => <ShowWrapperLoading />;
+
+export default AdminGalleryItemShowLoading;
