+++ src/app/admin/(modules)/gallery/[imageId]/page.tsx
@@ -1,4 +1,4 @@
-import ShowWrapper from "@/app/admin/gallery/[imageId]/_components/ShowWrapper";
+import ShowWrapper from "@/app/admin/(modules)/gallery/[imageId]/_components/ShowWrapper";
 import { Fragment } from "react";
 import { BreadcrumbDataSetter } from "@/app/admin/_components/Breadcrumbs";
 import { Metadata } from "next";
