+++ src/app/admin/(modules)/gallery/_components/RenameDialog.tsx
