+++ src/app/admin/(modules)/gallery/create/page.tsx
@@ -1,6 +1,6 @@
 import { Fragment } from "react";
 import { BreadcrumbDataSetter } from "@/app/admin/_components/Breadcrumbs";
-import GalleryCreateWrapper from "@/app/admin/gallery/create/_components/GalleryCreateWrapper";
+import GalleryCreateWrapper from "@/app/admin/(modules)/gallery/create/_components/GalleryCreateWrapper";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
