+++ src/app/admin/(modules)/gallery/page.tsx
