--- src/app/admin/gallery/[imageId]/loading.tsx
@@ -1,5 +1,0 @@
-import ShowWrapperLoading from "@/app/admin/gallery/[imageId]/_components/Loading/ShowWrapperLoading";
-
-const AdminGalleryItemShowLoading = () => <ShowWrapperLoading />;
-
-export default AdminGalleryItemShowLoading;
